requirements.txt
influxdb
paho-mqtt==1.6.1
